extract autoscroll hook
